[*] separate loop.hpp
